Fixed inconsistent connection handling.
If the app was not in memory, the LiveBluetoothSource wouldn't know there is an event to wait for as it never received the event itself, but was only launched in response to that.
The service would then subscribe to the list of devices which would never update as the event wasn't broadcast again.
In some edge cases it could still work if the device had connected completely by the point the initial LiveBluetoothSource data was loaded.

To fix the inconsistent behavior I moved the code back to a refresh based approach via BlueMusicService.java